Fix gtest dependencies
