+++ gclient/Cargo.toml
@@ -10,7 +10,7 @@ gear-utils.workspace = true
 gsdk = { workspace = true, features = ["testing"] }
 gear-core.workspace = true
 gear-core-errors.workspace = true
-gear-common = { workspace = true, default-features = true }
+gear-common.workspace = true
 
 futures.workspace = true
 anyhow.workspace = true
