+++ gclient/src/api/calls.rs
@@ -18,14 +18,12 @@
 
 use super::{GearApi, Result};
 use crate::{api::storage::account_id::IntoAccountId32, utils, Error};
-use gear_common::{
-    memory_dump::{MemoryPageDump, ProgramMemoryDump},
-    LockId,
-};
+use gear_common::LockId;
 use gear_core::{
     ids::*,
     memory::{GearPage, PageBuf, PageU32Size, GEAR_PAGE_SIZE, WASM_PAGE_SIZE},
 };
+use gear_utils::{MemoryPageDump, ProgramMemoryDump};
 use gsdk::{
     config::GearConfig,
     ext::{
